+++ django_tenants/staticfiles/storage.py
@@ -1,56 +1,49 @@
 import os
 
-from django.contrib.staticfiles.utils import check_settings
-from django.core.files.storage import FileSystemStorage
-
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.utils.functional import cached_property
 
 from django_tenants import utils
+from django_tenants.files.storage import TenantFileSystemStorage
 
 
-class TenantStaticFilesStorage(FileSystemStorage):
+class TenantStaticFilesStorage(TenantFileSystemStorage):
     """
     Implementation that extends core Django's StaticFilesStorage for multi-tenant setups.
+
+    The defaults for ``location`` and ``base_url`` are ``STATIC_ROOT`` and ``STATIC_URL``.
+
     """
+    def _clear_cached_properties(self, setting, **kwargs):
+        """Reset setting based property values."""
+        super()._clear_cached_properties(settings, **kwargs)
+
+        if setting == 'MULTITENANT_RELATIVE_STATIC_ROOT':
+            self.__dict__.pop('relative_static_root', None)
 
     @cached_property
     def relative_static_root(self):
         try:
-            return settings.MULTITENANT_RELATIVE_STATIC_ROOT
+            return os.path.join(settings.STATIC_ROOT, settings.MULTITENANT_RELATIVE_STATIC_ROOT)
         except AttributeError:
-            # MULTITENANT_RELATIVE_STATIC_ROOT is an optional setting, use the default value if none provided
-            # Use %s instead of "" to avoid raising exception every time in parse_tenant_config_path()
-            return "%s"
+            # MULTITENANT_RELATIVE_STATIC_ROOT is an optional setting. Use the default of just appending
+            # the tenant schema_name to STATIC_ROOT if no configuration value is provided
+            try:
+                return settings.STATIC_ROOT
+            except AttributeError:
+                raise ImproperlyConfigured("You're using the staticfiles app "
+                                           "without having set the STATIC_ROOT "
+                                           "setting to a filesystem path.")
 
-    @property
+    @property  # Not cached like in parent class
+    def base_location(self):
+        return utils.parse_tenant_config_path(self.relative_static_root)
+
+    @property  # Not cached like in parent class
     def base_url(self):
-        if self._base_url is not None and not self._base_url.endswith('/'):
-            self._base_url += '/'
-        _url = self._value_or_setting(self._base_url, settings.STATIC_URL)
+        url_ = settings.STATIC_URL
+        if not url_.endswith('/'):
+            url_ += '/'
 
-        _url = os.path.join(_url,
-                            utils.parse_tenant_config_path(self.relative_static_root))
-        if not _url.endswith("/"):
-            _url += "/"
-
-        check_settings(_url)
-
-        return _url
-
-    @property  # not cached like in parent class
-    def location(self):
-        _location = os.path.join(settings.STATIC_ROOT,
-                                 utils.parse_tenant_config_path(self.relative_static_root))
-        return os.path.abspath(_location)
-
-    def listdir(self, path):
-        """
-        More forgiving wrapper for parent class implementation that does not insist on
-        each tenant having its own static files dir.
-        """
-        try:
-            return super(TenantStaticFilesStorage, self).listdir(path)
-        except FileNotFoundError:
-            # Having static files for each tenant is optional - ignore.
-            return [], []
+        return url_