+++ django_tenants/tests/__init__.py
@@ -1,5 +1,3 @@
 from .test_routes import *
 from .test_tenants import *
 from .test_cache import *
-from .test_static_files import *
-from .test_filesystem_storage import *