+++ django_tenants/tests/template/loaders/__init__.py
@@ -1,0 +1,2 @@
+from .test_cached import *
+from .test_filesystem import *